fix: Get source output folder in lagacy Gradle.
- Call different APIs for different Gradle versions. Reflection
  is used because getOutputDir() does not exist in the current
  Gradle TAPIs.
