+++ testProjects/legacy-gradle/build.gradle
@@ -1,0 +1,10 @@
+plugins {
+	id 'java'
+}
+
+repositories {
+	mavenCentral()
+}
+
+dependencies {
+}
